refactor: rename ProviderConfigKey to ModelConfigKey, merge store files
Per Nir's review:
- Rename ProviderConfigKey → ModelConfigKey (config is model-scoped,
  provider config serves as defaults)
- Merge provider_store.go + model_store.go into store.go
