+++ pkg/plugins/model-provider-resolver/store_test.go
